pyproject.toml
[tool.poetry]
name = "navis_loader"
version = "0.1.0"
description = "Navis ETL Loader "
authors = ["Jason DeCorte <jdecorte@decorteindustries.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
pyspark = "^2.4"
python-dateutil = "^2.8"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pre-commit = "^1.18"

[tool.poetry.scripts]
naviload = 'navis_loader.cli:main'

[tool.black]
target_version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.idea
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
